+++ .storybook/preview.js
@@ -4,7 +4,14 @@ export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
   options: {
     storySort: {
-      order: ["Docs", ["Readme", "Changelog"], "Atoms", ["Button"]],
+      order: [
+        "Docs",
+        ["Readme", "Changelog"],
+        "Atoms",
+        ["Button"],
+        "Patterns",
+        ["Layout", ["Layout", "Navigation", "Main", "Footer"]],
+      ],
     },
   },
   controls: {