+++ src/components/Main/Main.tsx
@@ -10,10 +10,14 @@ export interface MainProps {
 
 const Main = (props: MainProps) => {
   return (
-    <div className="flex flex-0 items-center justify-center w-full bg-yellow-100 m-0 p-2.5">
+    <div className="flex flex-1 items-center min-h-screen justify-center w-full bg-beige-100 m-0 p-2.5">
       {props.children}
     </div>
   );
+};
+
+export const MainWrapper = (props: MainProps) => {
+  return <div className="flex flex-1 flex-col h-full">{props.children}</div>;
 };
 
 export default Main;