+++ src/components/Navigation/Navigation.tsx
@@ -10,7 +10,7 @@ export interface NavigationProps {
 
 const Navigation = (props: NavigationProps) => {
   return (
-    <div className="flex flex-col items-center justify-center w-1/5 m-2 px-4 bg-white">
+    <div className="flex flex-col items-center justify-center w-1/12 p-2 px-4 bg-white">
       <div className="sticky top-0 p-4 w-full h-full">{props.children}</div>
     </div>
   );