+++ src/components/index.ts
@@ -3,5 +3,5 @@ export { default as Button } from "./Button";
 // patterns
 export { default as Layout } from "./Layout";
 export { default as Navigation } from "./Navigation";
-export { default as Main } from "./Main";
+export { default as Main, MainWrapper as MainWrapper } from "./Main";
 export { default as Footer } from "./Footer";